requirements.txt
streamlit
tensorflow>=2.15
numpy
pillow
gtts
